backend/services/auth/requirements.txt
fastapi
uvicorn[standard]
redis
pydantic[email]
python-jose[cryptography]
python-multipart
aiosmtplib